fix: unify opportunity card layout across browse, related-rail and org-profile surfaces
The same opportunity rendered as two visually and informationally different
cards depending on where it was reached from: PublicOpportunityCard (org
profile "current needs", the detail page's "more from this organization"
rail) and OpportunityListItem (the /opportunities grid, the landing page
preview) diverged in title type scale, description colour, chip contents,
and - most importantly - the org-scoped cards dropped the date/deadline
entirely and showed only "One-time"/"Recurring" in its place. The top-right
chip slot also conflated three unrelated kinds of information (a spots-left
count, an "unlimited spots" flag, or a sign-up-mode pill) under one
inconsistent colour mapping.

Both components are replaced by one shared OpportunityCard, used by all four
surfaces with a headingLevel prop for heading-level consistency. The
top-right slot now always states one fact (sign-up mechanism via
participationType), moved out from under the capacity chip, which now sits
next to the date line instead - without duplicating "By expression of
interest" on interest-based cards, since the sign-up-mechanism chip is
omitted there rather than repeating what the capacity chip already says.

PublicOpportunitySummaryDto gained ValidUntil/NextTimeSlotStart so the
org-scoped cards can render a real date/deadline instead of dropping it -
the repository call behind it already resolved both, they were just never
carried onto this DTO.
