+++ frontend/AGENTS.md
@@ -15,7 +15,8 @@ src/
 │   ├── Header/             Header with auth state (login/logout buttons, org switcher, notifications) + optional per-page breadcrumb/action bar (see `breadcrumb` prop) - split across Header.tsx (orchestrator), DesktopHeader.tsx, MobileHeader.tsx, MobileMenu.tsx, AccountControls.tsx, NotificationDropdown.tsx, NotificationItem.tsx, OrganizationSwitcher.tsx, LanguageSelector.tsx, BreadcrumbBar.tsx - all exclusive to Header, nothing here is imported outside this folder
 │   ├── Footer.tsx          Footer with links and social icons
 │   ├── CreateOrganizationModal.tsx     Modal form for org creation
-│   ├── VolunteerOpportunitiesList/     Paginated list (size=10), filter bar - split across VolunteerOpportunitiesList.tsx (orchestrator), icons.tsx, MiniCalendar.tsx, FilterDropdown.tsx, OpportunityListItem.tsx, OpportunityResultsList.tsx, useVolunteerOpportunitiesData.ts, useOpportunityDateAvailability.ts, useCitySuggestions.ts - exclusive to this folder except `OpportunityListItem.tsx`, which `LatestOpportunitiesSection.tsx` (the landing page's three-card preview) also renders, so both surfaces show a visitor the same card
+│   ├── OpportunityCard.tsx     The one opportunity card, shared by every surface that shows one (#2054) - /opportunities, the landing page's three-card preview, the org profile's "current needs" list, and the opportunity detail page's "more from this organization" rail. Takes a `headingLevel` (2|3) prop and an `item` that either DTO (`VolunteerOpportunitySummary` or the leaner `PublicOpportunitySummaryDto`) satisfies structurally - fields absent on the leaner DTO (organization identity, tags, banner) gate the blocks that need them rather than switching components
+│   ├── VolunteerOpportunitiesList/     Paginated list (size=10), filter bar - split across VolunteerOpportunitiesList.tsx (orchestrator), icons.tsx, MiniCalendar.tsx, FilterDropdown.tsx, OpportunityResultsList.tsx, useVolunteerOpportunitiesData.ts, useOpportunityDateAvailability.ts, useCitySuggestions.ts - exclusive to this folder, all consuming the shared `OpportunityCard` above
 │   └── CreateVolunteerOpportunityModal.tsx  Modal form for opportunity creation
 ├── layouts/
 │   ├── AppLayout.tsx       Header + <Outlet /> + Footer
